Converted all doubles to float as requested

+++ src/main/Aufgabe6A.java
@@ -41,16 +41,16 @@ public class Aufgabe6A {
 		RangeCalculator rc = new RangeCalculator();
 		
 		RegressionCalculator rcSize = new RegressionCalculator(size);
-		double range_70 = rc.getRange(size, estimatedObjectLOC, 0.85);
-		double b0 = rcSize.getB0();
-		double b1 = rcSize.getB1();
-		double adjustedSize = b0 + estimatedObjectLOC * b1;
+		float range_70 = rc.getRange(size, estimatedObjectLOC, 0.85f);
+		float b0 = rcSize.getB0();
+		float b1 = rcSize.getB1();
+		float adjustedSize = b0 + estimatedObjectLOC * b1;
 		System.out.println("Size: ß0 = " + b0 + " | ß1 = " + b1);
 		System.out.println("Size-range(70%): " + range_70);
 		System.out.println("       UPI(70%): " + (adjustedSize + range_70));
 		System.out.println("       LPI(70%): " + (adjustedSize - range_70));
 		
-		double range_90 = rc.getRange(size, estimatedObjectLOC, 0.95);
+		float range_90 = rc.getRange(size, estimatedObjectLOC, 0.95f);
 		System.out.println("Size-range(90%): " + range_90);
 		System.out.println("       UPI(90%): " + (adjustedSize + range_90));
 		System.out.println("       LPI(90%): " + (adjustedSize - range_90));
@@ -66,8 +66,8 @@ public class Aufgabe6A {
 			}
 			RegressionCalculator rcTime = new RegressionCalculator(time);
 			int estimatedTime = (int) ((int) rcTime.getB0() + adjustedSize * rcTime.getB1());
-			double timeRange_70 = rc.getRange(time, estimatedTime, 0.85);
-			double timeRange_90 = rc.getRange(time, estimatedTime, 0.95);
+			float timeRange_70 = rc.getRange(time, estimatedTime, 0.85f);
+			float timeRange_90 = rc.getRange(time, estimatedTime, 0.95f);
 			System.out.println("Time: ß0 " + rcTime.getB0() + " | ß1 = " + rcTime.getB1());
 			System.out.println("Time-range(70%): = " + timeRange_70);
 			System.out.println("       UPI(70%): " + (estimatedTime + timeRange_70));
